pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["cr0un <vvmanikhin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pip = "^23.1.2"
python-dotenv = "^1.0.0"
pillow = "^9.5.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
django-cors-headers = "^3.14.0"
social-auth-app-django = "^5.2.0"
whitenoise = "^6.4.0"
gunicorn = "^20.1.0"
django-filter = "^23.2"
requests = "^2.30.0"
pydantic = "^1.10.7"
marshmallow = "^3.19.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'Todolist.settings'
python_files = [
    'tests.py',
    'test_*.py',
    '*_tests.py'
]
